error in the creation of mask, swapped columns/rows erroneously. Corrected this + added tons of stuff

+++ calculateAverageT2.m
@@ -4,7 +4,7 @@ function T2List = calculateAverageT2(T2Map,masks)
     T2List = zeros(1,masksSize(3));
     for i = 1:length(T2List)
         n = sum(sum(masks(:,:,i)));
-        T2List(i) = sum(sum(T2Map.*double(masks(:,:,i))))/n;
+        T2List(i) = sum(sum(T2Map.*(masks(:,:,i))))/n;
     end
     
 end
